react-security-scanner/requirements.txt
pydantic>=2.0.0
gitpython>=3.1.0
orchagent-sdk>=0.1.0
